requirements.txt
asgiref==3.11.0
colorama==0.4.6
coverage==7.12.0
dj-database-url==2.2.0
Django==5.2.8
gunicorn==22.0.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.2.1
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.7.0
sib-api-v3-sdk>=7.0.0,<8.0.0
requests>=2.31.0
